restructure `builder` into `dictionary`
