+++ crates/jpreprocess-dictionary/src/dictionary/to_dict/build_dict.rs
@@ -9,15 +9,13 @@ use rayon::prelude::*;
 use std::{collections::BTreeMap, str::FromStr};
 use yada::{builder::DoubleArrayBuilder, DoubleArray};
 
-use crate::DictionarySerializer;
+use crate::{dictionary::WordEntryMap, DictionarySerializer};
 
 const SIMPLE_USERDIC_FIELDS_NUM: usize = 3;
 const DETAILED_USERDIC_FIELDS_NUM: usize = 13;
 
 const SIMPLE_WORD_COST: i16 = -10000;
 const SIMPLE_CONTEXT_ID: u16 = 0;
-
-pub type WordEntryMap = BTreeMap<String, Vec<WordEntry>>;
 
 pub fn normalize_rows<'a, T, U, V>(rows: &'a T) -> Vec<Vec<String>>
 where
