+++ crates/jpreprocess-dictionary/src/dictionary/to_dict/mod.rs
@@ -1,3 +1,2 @@
 mod build_dict;
 pub mod ipadic_builder;
-pub mod to_csv;
